api refactor - finished moving crypto to separate trait
