+++ elrond-wasm-debug/src/api/crypto_api_mock.rs
@@ -18,4 +18,16 @@ impl CryptoApi for TxContext {
 		let hash: [u8; 32] = hasher.result().into();
 		hash.into()
 	}
+
+	fn verify_bls(&self, _key: &[u8], _message: &[u8], _signature: &[u8]) -> bool {
+		panic!("verify_bls not implemented yet!")
+	}
+
+	fn verify_ed25519(&self, _key: &[u8], _message: &[u8], _signature: &[u8]) -> bool {
+		panic!("verify_ed25519 not implemented yet!")
+	}
+
+	fn verify_secp256k1(&self, _key: &[u8], _message: &[u8], _signature: &[u8]) -> bool {
+		panic!("verify_secp256k1 not implemented yet!")
+	}
 }
