+++ elrond-wasm-debug/src/ext_mock.rs
@@ -346,34 +346,6 @@ impl elrond_wasm::ContractHookApi<RustBigInt, RustBigUint> for TxContext {
 		_function: &[u8], _arg_buffer: &ArgBuffer) {
 		panic!("execute_on_same_context not implemented yet!");
 	}
-
-	fn sha256(&self, data: &[u8]) -> H256 {
-		let mut hasher = Sha3_256::new();
-		hasher.input(data);
-		let hash: [u8; 32] = hasher.result().into();
-		hash.into()
-	}
-
-	fn keccak256(&self, data: &[u8]) -> H256 {
-		let mut hasher = Keccak256::new();
-		hasher.input(data);
-		let hash: [u8; 32] = hasher.result().into();
-		hash.into()
-	}
-
-	// TODO: Remove underscores when implementing
-
-	fn verify_bls(&self, _key: &[u8], _message: &[u8], _signature: &[u8]) -> bool {
-		panic!("verify_bls not implemented yet!")
-	}
-
-	fn verify_ed25519(&self, _key: &[u8], _message: &[u8], _signature: &[u8]) -> bool {
-		panic!("verify_ed25519 not implemented yet!")
-	}
-
-	fn verify_secp256k1(&self, _key: &[u8], _message: &[u8], _signature: &[u8]) -> bool {
-		panic!("verify_secp256k1 not implemented yet!")
-	}
 }
 
 impl elrond_wasm::ContractIOApi<RustBigInt, RustBigUint> for TxContext {
